+++ Sources/TSDocument-Syntax.m
@@ -27,9 +27,9 @@
 #import "globals.h"
 
 
-static BOOL isValidTeXCommandChar(int c);
+static BOOL isValidTeXCommandChar(NSInteger c);
 
-static BOOL isValidTeXCommandChar(int c)
+static BOOL isValidTeXCommandChar(NSInteger c)
 {
 	if ((c >= 'A') && (c <= 'Z'))
 		return YES;
@@ -77,13 +77,13 @@ static BOOL isValidTeXCommandChar(int c)
 {
 	NSLayoutManager *layoutManager;
 	NSString		*textString;
-	unsigned		length;
+	NSUInteger		length;
 	NSRange			colorRange;
-	unsigned		location;
-	int				theChar;
-	unsigned		aLineStart;
-	unsigned		aLineEnd;
-	unsigned		end;
+	NSUInteger		location;
+	NSInteger				theChar;
+	NSUInteger		aLineStart;
+	NSUInteger		aLineEnd;
+	NSUInteger		end;
 	BOOL			colorIndexDifferently;
 	NSTimeInterval	theTime;
 	
@@ -148,7 +148,7 @@ static BOOL isValidTeXCommandChar(int c)
 	
 	NSRange			lineRange, selectedLineRange, middleEastRange, testRange;
 	NSCharacterSet	*middleEastSet;
-	unsigned		start, theend;
+	NSUInteger		start, theend;
 	NSString		*theLine;
 	
 	if ([SUD boolForKey: RightJustifyKey]) {
@@ -164,18 +164,12 @@ static BOOL isValidTeXCommandChar(int c)
 			lineRange.location = theend;
 			selectedLineRange.location = start;
 			selectedLineRange.length = theend - start;
-			
-			if ( ! [SUD boolForKey: RightJustifyIfAnyKey] ) {
+            if ( ! [SUD boolForKey: RightJustifyIfAnyKey] ) {
                 // a line must START with Persian, etc., to be right justified; later must have Persian in first three letters
                 if (selectedLineRange.length >= 3)
                     selectedLineRange.length = 3;
-			}
-			
-	
-			// a line must START with Persian, etc., to be right justified
-			// if (selectedLineRange.length != 0)
-			// 	selectedLineRange.length = 1;
-			theLine = [textString substringWithRange:selectedLineRange];
+                }
+  			theLine = [textString substringWithRange:selectedLineRange];
 			testRange = [theLine rangeOfCharacterFromSet: middleEastSet];
 			if (testRange.location == NSNotFound)
 				[aTextView setAlignment: NSLeftTextAlignment range: selectedLineRange];
@@ -191,14 +185,14 @@ static BOOL isValidTeXCommandChar(int c)
 	colorRange.location = aLineStart;
 	colorRange.length = aLineEnd - aLineStart;
 	// WARNING!! The following line has been commented out to restore changing the text color
+    
 	// June 27, 2008; Koch; I don't understand the previous warning; the line below fixes cases when removing a comment leaves text red
-	// Sept 3, 2011; the Toudykov patch below makes this unnecessary
+    // Sept 3, 2011; the Toudykov patch below makes this unnecessary
 	// [layoutManager removeTemporaryAttribute:NSForegroundColorAttributeName forCharacterRange:colorRange];
     
     // the next line was added by Daniel Toundykov to allow changing the foreground and background source colors
     [layoutManager addTemporaryAttributes:regularColorAttribute forCharacterRange:colorRange];
     
-	
 	// Now we iterate over the whole text and perform the actual recoloring.
 	location = aLineStart;
 	while (location < aLineEnd) {
@@ -262,7 +256,7 @@ static BOOL isValidTeXCommandChar(int c)
 				NSString *commandString = [textString substringWithRange: colorRange];
 				// esindex below is a Spanish indexing command
 				if (([commandString isEqualToString: @"\\index"]) || ([commandString isEqualToString: @"\\esindex"])) {
-					int parens = 0;
+					NSInteger parens = 0;
 					BOOL optparens = NO;
 					BOOL notDone = YES;
 					
@@ -305,7 +299,7 @@ static BOOL isValidTeXCommandChar(int c)
 // Load the color definitions from the config system
 - (void)setupColors
 {
-	float		r, g, b;
+	CGFloat		r, g, b;
 	NSColor		*color;
 
 	//
@@ -422,10 +416,10 @@ static BOOL isValidTeXCommandChar(int c)
 
 
 // This is the main syntax coloring routine, used for everything except opening documents
-- (void)fixColor: (unsigned)from : (unsigned)to
+- (void)fixColor: (NSUInteger)from : (NSUInteger)to
 {
 	NSRange			colorRange;
-	unsigned		length;
+	NSUInteger		length;
 
 	// No syntax coloring if the file is not TeX, or if it is disabled
 	if (!fileIsTex || ![SUD boolForKey:SyntaxColoringEnabledKey])
