+++ Sources/TSStationery.h
@@ -9,9 +9,10 @@
 #import <Cocoa/Cocoa.h>
 
 #if __LP64__ || TARGET_OS_EMBEDDED || TARGET_OS_IPHONE || TARGET_OS_WIN32 || NS_BUILD_32_LIKE_64
+#warning 64BIT: Inspect use of long
 typedef long NSInteger;
 #else
-typedef int NSInteger;
+typedef NSInteger NSInteger;
 #endif
 
 
@@ -28,10 +29,10 @@ typedef int NSInteger;
 
 - (IBAction)newFromStationery: (id)sender;
 
-- (int)numberOfRowsInTableView:(NSTableView *)tableView;
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)tableView;
 - (id)tableView:(NSTableView *)tableView
 objectValueForTableColumn:(NSTableColumn *)tableColumn
-			row:(int)row;
+			row:(NSInteger)row;
 
 
 - (IBAction)okForStationeryPanel:(id)sender;
